+++ apps/meteor/app/livechat/server/methods/saveInfo.ts
@@ -7,7 +7,6 @@ import { Meteor } from 'meteor/meteor';
 import { callbacks } from '../../../../lib/callbacks';
 import { hasPermissionAsync } from '../../../authorization/server/functions/hasPermission';
 import { methodDeprecationLogger } from '../../../lib/server/lib/deprecationWarningLogger';
-import { Livechat } from '../lib/Livechat';
 import { Livechat as LivechatTyped } from '../lib/LivechatTyped';
 
 declare module '@rocket.chat/ui-contexts' {
@@ -78,9 +77,9 @@ Meteor.methods<ServerMethods>({
 			delete guestData.phone;
 		}
 
-		// Note: type is right, but `check` converts it to something else
 		// Since the endpoint is going to be deprecated, we can live with this for a while
-		await Promise.allSettled([LivechatTyped.saveGuest(guestData as any, userId), Livechat.saveRoomInfo(roomData)]);
+		// @ts-expect-error - type is right, but `check` converts it to something else
+		await Promise.allSettled([LivechatTyped.saveGuest(guestData as any, userId), LivechatTyped.saveRoomInfo(roomData)]);
 
 		const user = await Users.findOne({ _id: userId }, { projection: { _id: 1, username: 1 } });