+++ apps/meteor/app/utils/lib/i18n.ts
@@ -5,7 +5,7 @@ import { isObject } from '../../../lib/utils/isObject';
 
 export const i18n = i18next.use(sprintf);
 
-export const addSprinfToI18n = function (t: (typeof i18n)['t']) {
+export const addSprinfToI18n = function (t: (key: string, ...replaces: any) => string) {
 	return function (key: string, ...replaces: any): string {
 		if (replaces[0] === undefined || (isObject(replaces[0]) && !Array.isArray(replaces[0]))) {
 			return t(key, ...replaces);